browser entry
