+++ index.js
@@ -1,10 +1,10 @@
 module.exports = Object.assign(
   {
-    createBundler: require('./createBundler'),
     createClient: require('./createClient'),
     createDb: require('./createDb'),
     createLog: require('./createLog'),
-    createServer: require('./createServer'),
+    createApiServer: require('./createApiServer'),
+    createAssetServer: require('./createAssetServer'),
     createStore: require('./createStore'),
     Root: require('./Root')
   },
