+++ package.json
@@ -43,7 +43,7 @@
   },
   "dependencies": {
     "babel-register": "^6.24.1",
-    "bankai": "^9.11.1",
+    "bankai": "^9.12.0",
     "compression": "^1.6.2",
     "es2040": "^1.2.5",
     "express-enforces-ssl": "^1.1.0",
@@ -94,6 +94,7 @@
     "yargs": "^7.1.0"
   },
   "devDependencies": {
+    "ava": "^0.25.0",
     "dependency-check": "^2.7.0",
     "npm-run-all": "^4.0.1",
     "pg": "^6.2.2",
